requirements.txt
Django==2.2.10
djangorestframework==3.11.1
django-environ==0.4.5
psycopg2-binary==2.8.6
drf-yasg==1.17.1
